Fixing reassignment warning

+++ include/locate.h
@@ -1,3 +1,3 @@
 #include "global.h"
 
-void locate(float x, float *xx, int *j, int n);
+void locate(double x, double *xx, int *j, int n);
