+++ include/tabfind.h
@@ -1,6 +1,4 @@
 #include "global.h"
 #include "locate.h"
 
-
-
-void tabfind (double temperature, double *rate, float * atomic_temp_tab);
+void tabfind(double temperature, double *rate, double *atomic_temp_tab);
